Aligned alloc trying

+++ Assets/MemoryManager/MemoryUtil.cs
@@ -2,7 +2,7 @@ namespace Memory
 {
 	public class MemoryUtil
 	{
-		public static int AlignUp(int address, int alignment)
+		public static nint AlignUp(nint address, int alignment)
 		{
 			return address & ~(alignment - 1);
 		}
